Added Minhash Section

+++ MyAlgorithm/src/HashManager.java
@@ -3,6 +3,6 @@ public class HashManager {
 	public int[] hash_b;
 	public int prime_div;
 	public int HASH_SIZE; // the number of hashfunctions
-	public int[][] minhash_values;
-	public int[][] maxhash_values;
+	public long[][] minhash_values;
+	public long[][] maxhash_values;
 }
